Fix UTC Timestamp bugs
Helped to fix the timestamps
by refactoring the functions tweet.created_at
properties in the renderTwittlerStream and
the Tweet functions to print locale strings.

Also to note that the timeago function, which
evaluates each <time> node in the DOM is run
after each rendering of the twittler stream
(during initial boot up and when the refresh
button is hit), and after each submission
of a new tweet.