Finalize Twittler Project
Corrected the bugs present at
the moment, and polished the
applications.js with empathetic
code.  At the moment, I do not
have a unit tester, which in
hindsight, might have resolved
me of several headaches. At the
moment, content to a certain
degree with the final product,
since it satisfies the
requirements. Will revisit.